experiment/monitor/requirements.txt
requests~=2.31.0
kafka-python==2.0.2        # For interacting with Kafka